+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project does not currently adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2025.05.12
+
+### Added
+- Adding the gitstatus.sh script to the scripts section
+- Adding the gitstatus alias to zshrc to allow for easy running 
 
 ## 2025.05.09